styling fixes after merge
