+++ package.json
@@ -1,7 +1,7 @@
 {
 	"homepage": ".",
 	"name": "th2-infra-editor",
-	"version": "2.0.15",
+	"version": "2.0.16",
 	"private": true,
 	"dependencies": {
 		"@monaco-editor/react": "^4.2.1",
